pyproject.toml
[build-system]
requires      = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name        = "cifar10-mlops"
version     = "1.0.0"
description = "Multiclass image classification on CIFAR-10 with TensorFlow, MLflow, and Airflow"
requires-python = ">=3.10"
readme      = "README.md"
license     = {text = "MIT"}
authors     = [{name = "ML Engineer"}]

dependencies = [
    "tensorflow>=2.13.0",
    "mlflow>=2.9.0",
    "apache-airflow>=2.7.0",
    "numpy>=1.24.0",
    "pandas>=2.0.0",
    "scikit-learn>=1.3.0",
    "matplotlib>=3.7.0",
    "seaborn>=0.12.0",
    "PyYAML>=6.0",
    "tqdm>=4.66.0",
    "Pillow>=10.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=7.4.0", "pytest-cov>=4.1.0", "black", "ruff", "isort"]

[project.scripts]
train    = "scripts.run_training:main"
evaluate = "scripts.run_evaluation:main"
compare  = "scripts.compare_runs:main"

[tool.pytest.ini_options]
testpaths    = ["tests"]
addopts      = "-v --tb=short"
log_cli      = true
log_cli_level = "INFO"

[tool.coverage.run]
source = ["src"]
omit   = ["tests/*", "scripts/*", "dags/*"]

[tool.black]
line-length    = 95
target-version = ["py310", "py311", "py312"]

[tool.ruff]
line-length = 95
select      = ["E", "F", "W", "I"]

[tool.isort]
profile     = "black"
line_length = 95
